added need files

+++ database/migrations/2022_07_21_061224_create_specifications_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('specifications', function (Blueprint $table) {
             $table->id('specificationid');
             $table->string('description');
-            $table->unsignedInteger('productid');
+            $table->unsignedBigInteger('productid');
             $table->foreign('productid')->references('productid')->on('products');
         });
     }
